Throw when multiple configuration files are found
This mirrors new Babel 7 behavior. Collecting configuration from a
directory containing a `package.json` file now returns `null` if the
file does not encode an object, or if it has no `babel` field. If the
file does encode an object with a `babel` field, but that field is not
an object, it still returns an empty configuration.