Properly merge plugins and presets based on their name
When extending configurations, or using `env`, if an earlier plugin (or
preset, respectively) was declared with the same name its options are
replaced by that of a later plugin.

Mhen reducing configuration chains, if a plugin or preset was not
configured with a name, then a unique name is generated. Module paths
and exports are given the same name, to ensure that when plugins or
presets are merged the behavior is equivalent to Babel 7.